moving memstat.rs to vmm_tests
